requirements.txt
numpy
pandas
matplotlib
seaborn
tqdm
prettytable

torch
torchvision
torch_pruning
pytorch_quantization
thop

tensorrt
pycuda

opencv_python
pyts

wfdb
h5py
psutil
pyyaml
gdown